Globas.SCORE_COUNT
