--- src/main/java/cz/danik/Arctic.java
@@ -1,32 +1,0 @@
-package cz.danik;
-
-import cz.danik.graphics.Board;
-
-import javax.swing.*;
-import java.awt.*;
-
-public class Arctic extends JFrame {
-
-    public Arctic() {
-        initUI();
-    }
-
-    private void initUI() {
-        add(new Board());
-
-        setResizable(false);
-        pack();
-
-        setTitle("Arctic");
-        setLocationRelativeTo(null);
-        setDefaultCloseOperation(EXIT_ON_CLOSE);
-    }
-
-    public static void main(String[] args) {
-        EventQueue.invokeLater(() -> {
-            JFrame frame = new Arctic();
-            frame.setVisible(true);
-        });
-    }
-
-}
