+++ src/main/java/cz/danik/arctic/values/Globals.java
@@ -1,0 +1,14 @@
+package cz.danik.arctic.values;
+
+public class Globals {
+
+    public static int SCORE_COUNT;
+
+    public static int getScoreCount() {
+        return SCORE_COUNT;
+    }
+
+    public static void setScoreCount(int scoreCount) {
+        SCORE_COUNT = scoreCount;
+    }
+}
